requirement_latest.txt
asgiref==3.2.10
branca==0.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==37.0.2
defusedxml==0.7.1
Django==3.1.1
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-embed-video==1.4.1
folium==0.12.1.post1
gunicorn==20.0.4
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
oauthlib==3.2.0
paytmchecksum==1.7.0
Pillow==7.2.0
pycparser==2.21
pycryptodome==3.9.8
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.3.1
urllib3==1.26.8
whitenoise==5.2.0
